Library_Management_API/requirements.txt
Django>=4.1
djangorestframework>=3.14
